Create ProfileStyle.css
All the style elements I needed for this draft copy of the Profile page. It includes styles for both vertical and horizontal navigation bars, as well as styles for the drop down menu. Please note that the html also references the Bootstrap framework for additional style and layout of the Profile Page.